+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Last Changes
 
+- [#6](https://github.com/LaxarJS/laxar-infrastructure/issues/6): production: treat dependencies as externals
+
 
 ## v0.1.0-alpha.2